fix(compare): refactor compare to fix max depth error
