+++ frontend/src/views/WheresMyGene/components/Filters/components/Compare/types.ts
@@ -1,0 +1,3 @@
+export interface Props {
+  areFiltersDisabled: boolean;
+}
